Stash changes before rebase

+++ Tests/SwiftKueryPostgreSQLTests/TestInsert.swift
@@ -95,65 +95,89 @@ class TestInsert: XCTestCase {
                                         XCTAssertNotNil(rows, "INSERT returned no rows")
                                         let resultSet = result.asResultSet!
                                         XCTAssertEqual(rows!.count, 1, "INSERT returned wrong number of rows: \(rows!.count) instead of 1")
-                                        XCTAssertEqual(resultSet.titles[0], "a", "Wrong column name: \(resultSet.titles[0]) instead of a")
-                                        XCTAssertEqual(resultSet.titles[1], "b", "Wrong column name: \(resultSet.titles[1]) instead of b")
-                                        XCTAssertEqual(rows![0][0]! as! String, "apricot", "Wrong value in row 0 column 0")
-                                        XCTAssertEqual(rows![0][1]! as! Int32, 3, "Wrong value in row 1 column 0")
+                                        resultSet.getColumnTitles() { titles, error in
+                                            guard let titles = titles else {
+                                                XCTFail("No titles in result set")
+                                                return
+                                            }
+                                            XCTAssertEqual(titles[0], "a", "Wrong column name: \(titles[0]) instead of a")
+                                            XCTAssertEqual(titles[1], "b", "Wrong column name: \(titles[1]) instead of b")
+                                            XCTAssertEqual(rows![0][0]! as! String, "apricot", "Wrong value in row 0 column 0")
+                                            XCTAssertEqual(rows![0][1]! as! Int32, 3, "Wrong value in row 1 column 0")
 
-                                        let i3 = Insert(into: t, columns: [t.a, t.b], values: ["banana", 17])
-                                            .suffix("RETURNING b")
-                                        executeQuery(query: i3, connection: connection) { result, rows in
-                                            XCTAssertEqual(result.success, true, "INSERT failed")
-                                            XCTAssertNil(result.asError, "Error in INSERT: \(result.asError!)")
-                                            XCTAssertNotNil(result.asResultSet, "INSERT returned no rows")
-                                            XCTAssertNotNil(rows, "INSERT returned no rows")
-                                            let resultSet = result.asResultSet!
-                                            XCTAssertEqual(rows!.count, 1, "INSERT returned wrong number of rows: \(rows!.count) instead of 1")
-                                            XCTAssertEqual(resultSet.titles[0], "b", "Wrong column name: \(resultSet.titles[0]) instead of b")
-                                            XCTAssertEqual(resultSet.titles.count, 1, "Wrong number of columns: \(resultSet.titles.count) instead of 1")
-                                            XCTAssertEqual(rows![0][0]! as! Int32, 17, "Wrong value in row 0 column 0")
-
-                                            let i4 = Insert(into: t, rows: [["apple", 17], ["banana", -7], ["banana", 27]])
+                                            let i3 = Insert(into: t, columns: [t.a, t.b], values: ["banana", 17])
                                                 .suffix("RETURNING b")
-                                            executeQuery(query: i4, connection: connection) { result, rows in
+                                            executeQuery(query: i3, connection: connection) { result, rows in
                                                 XCTAssertEqual(result.success, true, "INSERT failed")
                                                 XCTAssertNil(result.asError, "Error in INSERT: \(result.asError!)")
                                                 XCTAssertNotNil(result.asResultSet, "INSERT returned no rows")
                                                 XCTAssertNotNil(rows, "INSERT returned no rows")
-                                                XCTAssertEqual(rows!.count, 3, "INSERT returned wrong number of rows: \(rows!.count) instead of 3")
+                                                let resultSet = result.asResultSet!
+                                                XCTAssertEqual(rows!.count, 1, "INSERT returned wrong number of rows: \(rows!.count) instead of 1")
+                                                resultSet.getColumnTitles() { titles, error in
+                                                    guard let titles = titles else {
+                                                    XCTFail("No titles in result set")
+                                                    return
+                                                    }
+                                                    XCTAssertEqual(titles[0], "b", "Wrong column name: \(titles[0]) instead of b")
+                                                    XCTAssertEqual(titles.count, 1, "Wrong number of columns: \(titles.count) instead of 1")
+                                                    XCTAssertEqual(rows![0][0]! as! Int32, 17, "Wrong value in row 0 column 0")
 
-                                                let i5 = Insert(into: t, rows: [["apple", 5], ["banana", 10], ["banana", 3]])
-                                                    .suffix("RETURNING b, a")
-                                                executeQuery(query: i5, connection: connection) { result, rows in
-                                                    XCTAssertEqual(result.success, true, "INSERT failed")
-                                                    XCTAssertNil(result.asError, "Error in INSERT: \(result.asError!)")
-                                                    XCTAssertNotNil(rows, "INSERT returned no rows")
-                                                    let resultSet = result.asResultSet!
-                                                    XCTAssertEqual(rows!.count, 3, "INSERT returned wrong number of rows: \(rows!.count) instead of 3")
-                                                    XCTAssertEqual(resultSet.titles.count, 2, "Wrong number of columns: \(resultSet.titles.count) instead of 2")
-
-                                                    let i6 = Insert(into: t2, Select(from: t).where(t.a == "apple"))
-                                                        .suffix("RETURNING *")
-                                                    executeQuery(query: i6, connection: connection) { result, rows in
+                                                    let i4 = Insert(into: t, rows: [["apple", 17], ["banana", -7], ["banana", 27]])
+                                                        .suffix("RETURNING b")
+                                                    executeQuery(query: i4, connection: connection) { result, rows in
                                                         XCTAssertEqual(result.success, true, "INSERT failed")
                                                         XCTAssertNil(result.asError, "Error in INSERT: \(result.asError!)")
+                                                        XCTAssertNotNil(result.asResultSet, "INSERT returned no rows")
                                                         XCTAssertNotNil(rows, "INSERT returned no rows")
-                                                        let resultSet = result.asResultSet!
                                                         XCTAssertEqual(rows!.count, 3, "INSERT returned wrong number of rows: \(rows!.count) instead of 3")
-                                                        XCTAssertEqual(resultSet.titles.count, 2, "Wrong number of columns: \(resultSet.titles.count) instead of 2")
 
-                                                        let s1 = Select(from: t)
-                                                        executeQuery(query: s1, connection: connection) { result, rows in
-                                                            XCTAssertEqual(result.success, true, "SELECT failed")
-                                                            XCTAssertNil(result.asError, "Error in SELECT: \(result.asError!)")
-                                                            XCTAssertNotNil(rows, "SELECT returned no rows")
-                                                            XCTAssertEqual(rows!.count, 9, "INSERT returned wrong number of rows: \(rows!.count) instead of 9")
+                                                        let i5 = Insert(into: t, rows: [["apple", 5], ["banana", 10], ["banana", 3]])
+                                                            .suffix("RETURNING b, a")
+                                                        executeQuery(query: i5, connection: connection) { result, rows in
+                                                            XCTAssertEqual(result.success, true, "INSERT failed")
+                                                            XCTAssertNil(result.asError, "Error in INSERT: \(result.asError!)")
+                                                            XCTAssertNotNil(result.asResultSet, "INSERT returned no rows")
+                                                            XCTAssertNotNil(rows, "INSERT returned no rows")
+                                                            XCTAssertEqual(rows!.count, 3, "INSERT returned wrong number of rows: \(rows!.count) instead of 3")
+                                                            resultSet.getColumnTitles() { titles, error in
+                                                                guard let titles = titles else {
+                                                                    XCTFail("No titles in result set")
+                                                                    return
+                                                                }
+                                                                XCTAssertEqual(titles.count, 2, "Wrong number of columns: \(titles.count) instead of 2")
 
-                                                            let dropT = Raw(query: "DROP TABLE", table: t)
-                                                            executeQuery(query: dropT, connection: connection) { result, rows in
-                                                                XCTAssertEqual(result.success, true, "DROP TABLE failed")
-                                                                XCTAssertNil(result.asError, "Error in DELETE: \(result.asError!)")
-                                                                expectation.fulfill()
+                                                                let i6 = Insert(into: t2, Select(from: t).where(t.a == "apple"))
+                                                                    .suffix("RETURNING *")
+                                                                executeQuery(query: i6, connection: connection) { result, rows in
+                                                                    XCTAssertEqual(result.success, true, "INSERT failed")
+                                                                    XCTAssertNil(result.asError, "Error in INSERT: \(result.asError!)")
+                                                                    XCTAssertNotNil(rows, "INSERT returned no rows")
+                                                                    let resultSet = result.asResultSet!
+                                                                    XCTAssertEqual(rows!.count, 3, "INSERT returned wrong number of rows: \(rows!.count) instead of 3")
+                                                                    resultSet.getColumnTitles() { titles, error in
+                                                                        guard let titles = titles else {
+                                                                            XCTFail("No titles in result set")
+                                                                            return
+                                                                        }
+                                                                        XCTAssertEqual(titles.count, 2, "Wrong number of columns: \(titles.count) instead of 2")
+
+                                                                        let s1 = Select(from: t)
+                                                                        executeQuery(query: s1, connection: connection) { result, rows in
+                                                                            XCTAssertEqual(result.success, true, "SELECT failed")
+                                                                            XCTAssertNil(result.asError, "Error in SELECT: \(result.asError!)")
+                                                                            XCTAssertNotNil(rows, "SELECT returned no rows")
+                                                                            XCTAssertEqual(rows!.count, 9, "INSERT returned wrong number of rows: \(rows!.count) instead of 9")
+
+                                                                            let dropT = Raw(query: "DROP TABLE", table: t)
+                                                                            executeQuery(query: dropT, connection: connection) { result, rows in
+                                                                                XCTAssertEqual(result.success, true, "DROP TABLE failed")
+                                                                                XCTAssertNil(result.asError, "Error in DELETE: \(result.asError!)")
+                                                                                expectation.fulfill()
+                                                                            }
+                                                                        }
+                                                                    }
+                                                                }
                                                             }
                                                         }
                                                     }
@@ -193,13 +217,19 @@ class TestInsert: XCTestCase {
                             XCTAssertNotNil(rows, "INSERT returned no rows")
                             let resultSet = result.asResultSet!
                             XCTAssertEqual(rows!.count, 1, "INSERT returned wrong number of rows: \(rows!.count) instead of 1")
-                            XCTAssertEqual(resultSet.titles.count, 1, "Wrong number of columns: \(resultSet.titles.count) instead of 1")
+                            resultSet.getColumnTitles() { titles, error in
+                                guard let titles = titles else {
+                                    XCTFail("No titles in result set")
+                                    return
+                                }
+                                XCTAssertEqual(titles.count, 1, "Wrong number of columns: \(titles.count) instead of 1")
 
-                            let dropT3 = Raw(query: "DROP TABLE", table: t3)
-                            executeQuery(query: dropT3, connection: connection) { result, rows in
-                                XCTAssertEqual(result.success, true, "DROP TABLE failed")
-                                XCTAssertNil(result.asError, "Error in DELETE: \(result.asError!)")
-                                expectation.fulfill()
+                                let dropT3 = Raw(query: "DROP TABLE", table: t3)
+                                executeQuery(query: dropT3, connection: connection) { result, rows in
+                                    XCTAssertEqual(result.success, true, "DROP TABLE failed")
+                                    XCTAssertNil(result.asError, "Error in DELETE: \(result.asError!)")
+                                    expectation.fulfill()
+                                }
                             }
                         }
                     }
